+++ src/constants.ts
@@ -3,4 +3,6 @@ export const DEFAULT_SOURCE_DIRECTORY_NAME = 'src';
 export const DEFAULT_ASSETS_DIRECTORY_NAME = 'assets';
 export const DEFAULT_IMAGES_DIRECTORY_NAME = 'images';
 
-export const SVG_CLASS_REGEXP = /<svg\s+[^>]*class=['"]([^'"]+)['"]/;
+export const SVG_NAME = 'svg';
+
+export const URL_DELIMITER = '/';